devide sub_module
