+++ music/walk_by_val/main.py
@@ -1,40 +1,18 @@
-
 # coding: utf-8
-
-# # 構成
-# - 画像を音楽に変換するプログラム. 
-# - defの集まり > main という構成. 
-# - 音楽制作はmelodyとloop(の繰り返し)の2種類で構成している. 
-
 
 import numpy as np
 import cv2
-import matplotlib.pyplot as plt
-from itertools import chain, zip_longest
 import random
 import time
-import math
-import wave
-import struct
-import random
-from definitions import partition1d, partition2d, to_mean, hsv2rgb, rgb2hsv
-from definitions import make_all_candidate, dont_back, select_available, decide_next, decide_shift, jump_unknown
-from definitions import mk_reduced_color_img, normalize1d, path_array
-from definitions import mk_reduced_rgb, mk_reduced_hsv
-from definitions import img2path
-from definitions import makewave, wavwrite
-from definitions import ampFn, waveFn, set_key, value2freq, unify1d, value2first_time, value2time_length
-from definitions import path2wave
-
+import image
+import music
 
 start_whole_time = time.time()
 sr = 44100 # sampling rate
 
-
 img = cv2.imread("/Users/koichi-sakaguchi/prv/private/chaofan.jpg")
 img_name = "picasso"
 img = cv2.cvtColor(img,cv2.COLOR_BGR2RGB)
-
 
 partition_num_width = 10
 partition_num_height = 10
@@ -48,29 +26,22 @@ if is_loop == 0 :
 elif is_loop == 1 :
     mode_name = "loop"
 
-log_id, reduced_rgb, reduced_hsv = img2path(img, partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num)
+log_id, reduced_rgb, reduced_hsv = image.img2path(img, partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num)
 
 
 print(log_id)
 
 img_size = 400 # pixel
-reduced_color_img = mk_reduced_color_img(img_size, partition_num_width, partition_num_height, log_id, reduced_rgb, mode_name)
-plt.imshow(reduced_color_img)
+reduced_color_img = image.mk_reduced_color_img(img_size, partition_num_width, partition_num_height, log_id, reduced_rgb, mode_name)
 
-path_h = path_array(log_id, reduced_hsv[:,:,0])
-path_s = path_array(log_id, reduced_hsv[:,:,1])
-path_v = path_array(log_id, reduced_hsv[:,:,2])
+path_h = image.path_array(log_id, reduced_hsv[:,:,0])
+path_s = image.path_array(log_id, reduced_hsv[:,:,1])
+path_v = image.path_array(log_id, reduced_hsv[:,:,2])
 
-path_normalized_h = normalize1d(path_h)
-path_normalized_s = normalize1d(path_s)
-path_normalized_v = normalize1d(path_v)
+path_normalized_h = image.normalize1d(path_h)
+path_normalized_s = image.normalize1d(path_s)
+path_normalized_v = image.normalize1d(path_v)
 path_normalized_hsv = [path_normalized_h, path_normalized_s, path_normalized_v]
-
-fig = plt.figure(figsize=(20,5))
-ax1, ax2, ax3 = fig.add_subplot(131), fig.add_subplot(132), fig.add_subplot(133)
-ax1.plot(path_h)
-ax2.plot(path_s)
-ax3.plot(path_v)
 
 is_cut_loop = 0
 cut_loop_num = 1
@@ -91,7 +62,7 @@ time_1count    = 1/bps
 time_element   = time_1count/2
 time_1loop       = loop_count*time_1count
 
-unified_sound_wave, path_freq, path_first_time, path_time_length = path2wave(path_normalized_hsv, keyname, first_octave, plus_octave, key_id_shift, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, loop_count, len_id_usage, sr)
+unified_sound_wave, path_freq, path_first_time, path_time_length = music.path2wave(path_normalized_hsv, keyname, first_octave, plus_octave, key_id_shift, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, loop_count, len_id_usage, sr)
 if is_cut_loop == 1:
     len_1loop = int(cut_loop_num*time_1loop*sr)
     unified_sound_wave = unified_sound_wave[0:len_1loop]
@@ -103,22 +74,13 @@ raw_unified_sound_wave = unified_sound_wave
 
 whole_time_length = len(unified_sound_wave)/sr
 
-
-fig = plt.figure(figsize=(20,5))
-ax1, ax2, ax3, ax4 = fig.add_subplot(141), fig.add_subplot(142), fig.add_subplot(143), fig.add_subplot(144)
-
-ax1.plot(unified_sound_wave)
-ax2.plot(path_first_time)
-ax3.plot(path_time_length)
-ax4.plot(path_freq)
-
 if is_loop == 1:
     unified_sound_wave = np.array(int(len(unified_sound_wave_melody)/len(raw_unified_sound_wave) + 1)*list(raw_unified_sound_wave))
     unified_sound_wave_loop = unified_sound_wave    
 
 max_amp = max(abs(unified_sound_wave))
 unified_sound_wave = unified_sound_wave/max_amp
-wavwrite(unified_sound_wave,"music_w"+str(partition_num_width)+"h"+str(partition_num_height)+".wav")
+music.wavwrite(unified_sound_wave,"music_w"+str(partition_num_width)+"h"+str(partition_num_height)+".wav")
 
 img_prm = [partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num]
 msc_prm = [is_cut_loop, cut_loop_num, is_loop, keyname, first_octave, plus_octave, key_id_shift, loop_count, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, len_id_usage]
@@ -145,29 +107,19 @@ if is_loop == 0 :
 elif is_loop == 1 :
     mode_name = "loop"
 
-log_id, reduced_rgb, reduced_hsv = img2path(img, partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num)
-
-print(log_id)
-print(len(log_id))
+log_id, reduced_rgb, reduced_hsv = image.img2path(img, partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num)
 
 img_size = 400 # pixel
-reduced_color_img = mk_reduced_color_img(img_size, partition_num_width, partition_num_height, log_id, reduced_rgb, mode_name)
-plt.imshow(reduced_color_img)
+reduced_color_img = image.mk_reduced_color_img(img_size, partition_num_width, partition_num_height, log_id, reduced_rgb, mode_name)
 
-path_h = path_array(log_id, reduced_hsv[:,:,0])
-path_s = path_array(log_id, reduced_hsv[:,:,1])
-path_v = path_array(log_id, reduced_hsv[:,:,2])
+path_h = image.path_array(log_id, reduced_hsv[:,:,0])
+path_s = image.path_array(log_id, reduced_hsv[:,:,1])
+path_v = image.path_array(log_id, reduced_hsv[:,:,2])
 
-path_normalized_h = normalize1d(path_h)
-path_normalized_s = normalize1d(path_s)
-path_normalized_v = normalize1d(path_v)
+path_normalized_h = image.normalize1d(path_h)
+path_normalized_s = image.normalize1d(path_s)
+path_normalized_v = image.normalize1d(path_v)
 path_normalized_hsv = [path_normalized_h, path_normalized_s, path_normalized_v]
-
-fig = plt.figure(figsize=(20,5))
-ax1, ax2, ax3 = fig.add_subplot(131), fig.add_subplot(132), fig.add_subplot(133)
-ax1.plot(path_h)
-ax2.plot(path_s)
-ax3.plot(path_v)
 
 is_cut_loop = 1
 cut_loop_num = 1
@@ -188,7 +140,7 @@ time_1count    = 1/bps
 time_element   = time_1count/2
 time_1loop       = loop_count*time_1count
 
-unified_sound_wave, path_freq, path_first_time, path_time_length = path2wave(path_normalized_hsv, keyname, first_octave, plus_octave, key_id_shift, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, loop_count, len_id_usage, sr)
+unified_sound_wave, path_freq, path_first_time, path_time_length = music.path2wave(path_normalized_hsv, keyname, first_octave, plus_octave, key_id_shift, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, loop_count, len_id_usage, sr)
 whole_time_length = len(unified_sound_wave)/sr
 
 if is_cut_loop == 1:
@@ -202,14 +154,6 @@ raw_unified_sound_wave = unified_sound_wave
 
 whole_time_length = len(unified_sound_wave)/sr
 
-fig = plt.figure(figsize=(20,5))
-ax1, ax2, ax3, ax4 = fig.add_subplot(141), fig.add_subplot(142), fig.add_subplot(143), fig.add_subplot(144)
-
-ax1.plot(unified_sound_wave)
-ax2.plot(path_first_time)
-ax3.plot(path_time_length)
-ax4.plot(path_freq)
-
 if is_loop == 1:
     looped_unified_sound_wave = np.array(int(len(unified_sound_wave_melody)/len(raw_unified_sound_wave) + 1)*list(raw_unified_sound_wave))
     unified_sound_wave = looped_unified_sound_wave
@@ -217,7 +161,7 @@ if is_loop == 1:
 max_amp = max(abs(unified_sound_wave))
 print(max_amp)
 unified_sound_wave = unified_sound_wave/max_amp
-wavwrite(unified_sound_wave,"music_w"+str(partition_num_width)+"h"+str(partition_num_height)+".wav")
+music.wavwrite(unified_sound_wave,"music_w"+str(partition_num_width)+"h"+str(partition_num_height)+".wav")
 
 img_prm = [partition_num_width, partition_num_height, last_path_order, first_shift, repeatable_num]
 msc_prm = [is_cut_loop, cut_loop_num, is_loop, keyname, first_octave, plus_octave, key_id_shift, loop_count, first_time_ratio, time_length_ratio, isnt_degeneratable, bpm, len_id_usage]
@@ -233,12 +177,12 @@ unified_sound_wave_loop = looped_unified_sound_wave
 partition_num_width_loop = partition_num_width
 partition_num_height_loop = partition_num_height
 
-unified_sound_wave = unify1d(1*unified_sound_wave_melody, 1*unified_sound_wave_loop)
+unified_sound_wave = music.unify1d(1*unified_sound_wave_melody, 1*unified_sound_wave_loop)
 
 max_amp = max(abs(unified_sound_wave))
 print(max_amp)
 unified_sound_wave = unified_sound_wave/max_amp
-wavwrite(unified_sound_wave,"music_"+prm_mldy+"_"+prm_loop+"_"+img_name+".wav")
+music.wavwrite(unified_sound_wave,"music_"+prm_mldy+"_"+prm_loop+"_"+img_name+".wav")
 
 end_whole_time = time.time()
 elapsed_whole_time = end_whole_time - start_whole_time
